move paper wallet from settings to create wallet
