+++ app/components/wallet/add/paper-wallets/CreatePaperDialog.js
