+++ app/components/wallet/add/paper-wallets/CreatePaperDialog.scss
