+++ app/components/wallet/add/paper-wallets/FinalizeDialog.scss
