+++ app/components/wallet/add/paper-wallets/LoadingGif.scss
