+++ app/components/wallet/add/paper-wallets/UserPasswordDialog.js
