+++ app/containers/wallet/dialogs/CreatePaperWalletDialogContainer.js
@@ -3,16 +3,16 @@ import type { Node } from 'react';
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import { computed, observable, runInAction, } from 'mobx';
-import UserPasswordDialog from '../../../components/wallet/settings/paper-wallets/UserPasswordDialog';
+import UserPasswordDialog from '../../../components/wallet/add/paper-wallets/UserPasswordDialog';
 import type { InjectedOrGenerated } from '../../../types/injectedPropsType';
 import config from '../../../config';
 import { ProgressStep } from '../../../stores/ada/PaperWalletCreateStore';
 import { Logger } from '../../../utils/logging';
-import CreatePaperDialog from '../../../components/wallet/settings/paper-wallets/CreatePaperDialog';
-import LoadingGif from '../../../components/wallet/settings/paper-wallets/LoadingGif';
+import CreatePaperDialog from '../../../components/wallet/add/paper-wallets/CreatePaperDialog';
+import LoadingGif from '../../../components/wallet/add/paper-wallets/LoadingGif';
 import WalletRestoreDialog from '../../../components/wallet/WalletRestoreDialog';
 import validWords from 'bip39/src/wordlists/english.json';
-import FinalizeDialog from '../../../components/wallet/settings/paper-wallets/FinalizeDialog';
+import FinalizeDialog from '../../../components/wallet/add/paper-wallets/FinalizeDialog';
 import type { AdaPaper } from '../../../api/ada';
 import { defineMessages, intlShape } from 'react-intl';
 import globalMessages from '../../../i18n/global-messages';
