+++ app/i18n/locales/en-US.json
@@ -208,7 +208,6 @@
   "settings.menu.externalStorage.link.label": "Memo Storage",
   "settings.menu.general.link.label": "General",
   "settings.menu.levelOfComplexity.link.label": "Level of Complexity",
-  "settings.menu.paperWallet.link.label": "Paper Wallet",
   "settings.menu.support.link.label": "Support / Logs",
   "settings.menu.termsOfUse.link.label": "Terms of use",
   "settings.menu.wallet.link.label": "Wallet",
@@ -319,6 +318,8 @@
   "wallet.add.optionDialog.connect.hw.ledger.title": "Ledger Hardware Wallet",
   "wallet.add.optionDialog.connect.hw.trezor.learnMoreText": "A Trezor hardware wallet is a small USB device that adds an extra level of security to your wallet. It is more secure because your private key never leaves the hardware wallet. This protects your funds even if your computer is compromised due to malware, phishing attempts, etc.",
   "wallet.add.optionDialog.connect.hw.trezor.title": "Trezor Hardware Wallet",
+  "wallet.add.optionDialog.create.normalWallet.description": "A standard wallet backed by a recovery phrase.",
+  "wallet.add.optionDialog.create.paperWallet.description": "Paper wallets can be created even on devices not connected to the internet which makes them well-suited for single-use cold storage.",
   "wallet.add.optionDialog.restore.dialogTitle": "Restore wallet",
   "wallet.add.optionDialog.restore.normalWallet.description": "If you have a Yoroi recovery phrase consisting of 15 words generated when you created a Yoroi Wallet, choose this option to restore your wallet.",
   "wallet.add.optionDialog.restore.normalWallet.title": "Enter a 15-word recovery phrase",
