+++ features/wallet-creation.feature
@@ -9,6 +9,7 @@ Feature: Wallet creation
   Scenario: Wallet creation (IT-5)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password   | repeatedPassword  |
@@ -22,6 +23,7 @@ Feature: Wallet creation
   Scenario: Wallet access after browser restart (IT-9)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password   | repeatedPassword  |
@@ -37,6 +39,7 @@ Feature: Wallet creation
   Scenario: Mnemonic words can be cleared by clicking "Clear button" on wallet creation screen (IT-18)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password   | repeatedPassword  |
@@ -51,6 +54,7 @@ Feature: Wallet creation
   Scenario: Wallet can't be created without entering password (IT-24)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password   | repeatedPassword  |
@@ -65,6 +69,7 @@ Feature: Wallet creation
    Scenario: Users will be presented with a security warning prior to seed creation (IT-27)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password   | repeatedPassword  |
@@ -78,6 +83,7 @@ Feature: Wallet creation
   Scenario Outline: Wallet can't be created if wallet name doesn't meet requirements (IT-16)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password   | repeatedPassword  |
@@ -100,6 +106,7 @@ Feature: Wallet creation
   Scenario Outline: Wallet can't be created if its password doesn't meet complexity requirements (IT-7)
     When I click the create button
     Then I select Cardano
+    Then I select Create Wallet
     And I enter the name "Created Wallet"
     And I enter the created wallet password:
     | password        | repeatedPassword  |
